zos: correct epoll behaviour for invalid fds
The z/OS epoll implementation needs to perform certain checks
on file descriptors and return ENOENT or EINVAL if those fds
aren't tracked or invalid (e.g. -1).

+++ src/unix/os390-syscalls.h
@@ -49,7 +49,7 @@ struct epoll_event {
 typedef struct {
   QUEUE member;
   struct pollfd* items;
-  unsigned long size;
+  int size;
   int msg_queue;
 } uv__os390_epoll;
